update to .net 10 preview, update lazer packages, fix maxcombo being a double

+++ osu.Native/Structures/Difficulty/NativeCatchDifficultyAttributes.cs
@@ -5,5 +5,5 @@ namespace osu.Native.Structures.Difficulty;
 internal struct NativeCatchDifficultyAttributes(CatchDifficultyAttributes attributes)
 {
   public double StarRating = attributes.StarRating;
-  public double MaxCombo = attributes.MaxCombo;
+  public int MaxCombo = attributes.MaxCombo;
 }
