+++ osu.Native/Structures/Difficulty/NativeOsuDifficultyAttributes.cs
@@ -5,7 +5,7 @@ namespace osu.Native.Structures.Difficulty;
 internal struct NativeOsuDifficultyAttributes(OsuDifficultyAttributes attributes)
 {
   public double StarRating = attributes.StarRating;
-  public double MaxCombo = attributes.MaxCombo;
+  public int MaxCombo = attributes.MaxCombo;
   public double AimDifficulty = attributes.AimDifficulty;
   public double AimDifficultSliderCount = attributes.AimDifficultSliderCount;
   public double SpeedDifficulty = attributes.SpeedDifficulty;
